Split cache acceptance tests
